Gameboy/Dot_Matrix video filters: Add XRGB8888 support

+++ gfx/video_filters/scanline2x.c
@@ -108,11 +108,11 @@ static void scanline2x_generic_destroy(void *data)
 static void scanline2x_work_cb_xrgb8888(void *data, void *thread_data)
 {
    struct softfilter_thread_data *thr = (struct softfilter_thread_data*)thread_data;
-   const uint32_t *input = (const uint32_t*)thr->in_data;
-   uint32_t *output = (uint32_t*)thr->out_data;
-   unsigned in_stride = (unsigned)(thr->in_pitch >> 2);
-   unsigned out_stride = (unsigned)(thr->out_pitch >> 2);
-   unsigned x, y;
+   const uint32_t *input              = (const uint32_t*)thr->in_data;
+   uint32_t *output                   = (uint32_t*)thr->out_data;
+   uint32_t in_stride                 = (uint32_t)(thr->in_pitch >> 2);
+   uint32_t out_stride                = (uint32_t)(thr->out_pitch >> 2);
+   uint32_t x, y;
 
    for (y = 0; y < thr->height; ++y)
    {
@@ -124,10 +124,10 @@ static void scanline2x_work_cb_xrgb8888(void *data, void *thread_data)
           * byte swapping issues */
          uint32_t *out_line_ptr  = out_ptr;
          uint32_t color          = *(input + x);
-         uint8_t  p              = (color >> 24 & 0xFF); /* Padding bits */
-         uint8_t  r              = (color >> 16 & 0xFF);
-         uint8_t  g              = (color >>  8 & 0xFF);
-         uint8_t  b              = (color       & 0xFF);
+         uint32_t  p             = (color >> 24 & 0xFF); /* Padding bits */
+         uint32_t  r             = (color >> 16 & 0xFF);
+         uint32_t  g             = (color >>  8 & 0xFF);
+         uint32_t  b             = (color       & 0xFF);
          uint32_t scanline_color =
                ((p - (p >> 2)) << 24) |
                ((r - (r >> 2)) << 16) |
@@ -154,11 +154,11 @@ static void scanline2x_work_cb_xrgb8888(void *data, void *thread_data)
 static void scanline2x_work_cb_rgb565(void *data, void *thread_data)
 {
    struct softfilter_thread_data *thr = (struct softfilter_thread_data*)thread_data;
-   const uint16_t *input = (const uint16_t*)thr->in_data;
-   uint16_t *output = (uint16_t*)thr->out_data;
-   unsigned in_stride = (unsigned)(thr->in_pitch >> 1);
-   unsigned out_stride = (unsigned)(thr->out_pitch >> 1);
-   unsigned x, y;
+   const uint16_t *input              = (const uint16_t*)thr->in_data;
+   uint16_t *output                   = (uint16_t*)thr->out_data;
+   uint16_t in_stride                 = (uint16_t)(thr->in_pitch >> 1);
+   uint16_t out_stride                = (uint16_t)(thr->out_pitch >> 1);
+   uint16_t x, y;
 
    for (y = 0; y < thr->height; ++y)
    {
@@ -167,9 +167,9 @@ static void scanline2x_work_cb_rgb565(void *data, void *thread_data)
       {
          uint16_t *out_line_ptr  = out_ptr;
          uint16_t color          = *(input + x);
-         uint8_t  r              = (color >> 11 & 0x1F);
-         uint8_t  g              = (color >>  6 & 0x1F);
-         uint8_t  b              = (color       & 0x1F);
+         uint16_t r              = (color >> 11 & 0x1F);
+         uint16_t g              = (color >>  6 & 0x1F);
+         uint16_t b              = (color       & 0x1F);
          uint16_t scanline_color =
                ((r - (r >> 2)) << 11) |
                ((g - (g >> 2)) <<  6) |
